Add loading of TGMPA native translation files.
Implemented in a way that the language loading will work independently of whether TGMPA is included in a plugin or a theme.
Including compatibility for looking in both the WP_LANG_DIR/plugins/ as well as the WP_LANG_DIR/themes/ directory for translations served by Wp.org.

Ignore warnings from PHPCS as we *really* should suppress warning on `is_readable()`.